+++ src/main/java/com/alipay/global/api/response/ams/billing/AlipayPriceUpdateResponse.java
@@ -55,7 +55,7 @@ public class AlipayPriceUpdateResponse extends AlipayResponse {
   private Boolean active;
 
   /** The included quantity. */
-  private Integer includedQuantity;
+  private Long includedQuantity;
 
   /** The tiers mode. Maximum length: 16 characters. */
   private String tiersMode;
